Further refactoring and moving of files.
