--- src/actions/index.tsx
@@ -1,9 +1,0 @@
-export const addMessage = (message: any) => {
-  return {
-    type: 'ADD_MESSAGE',
-    id: 0,
-    message,
-  };
-};
-
-// TODO: ADD ACTIONS HERE
