+++ src/components/VisualCourse/course.tsx
