+++ src/components/VisualCourse/displayedCourse.css
