+++ src/components/VisualCourse/displayedCourse.tsx
