+++ src/pages/Browse.tsx
@@ -20,7 +20,7 @@ import { connect } from 'react-redux';
 
 import { changeFilters, submitSearch } from '../redux/actions/Browse';
 import { Filters, Props } from '../type-interfaces/Browse'; // eslint-disable-line no-unused-vars
-import { MIN_COURSE_CODE, MAX_COURSE_CODE, SLIDER_STEP_SIZE } from './UIConstants';
+import { MIN_COURSE_CODE, MAX_COURSE_CODE, SLIDER_STEP_SIZE } from '../util/UIConstants';
 
 const useStyles = makeStyles((theme: Theme) => createStyles({
   root: {
