+++ src/redux/actions/browse.tsx
@@ -1,7 +1,10 @@
+import { CHANGE_FILTERS, SUBMIT_SEARCH } from '../constants';
+import { Filters } from '../../type-interfaces/Browse'; // eslint-disable-line no-unused-vars
+
 // eslint-disable-next-line arrow-body-style
-export const submitSearch = (filters: { query: string, letterCodes: Array<string>, numberRange: Array<number> }) => {
+export const changeFilters = (filters: Filters) => {
   return {
-    type: 'SUBMIT_SEARCH',
+    type: CHANGE_FILTERS,
     query: filters.query,
     letterCodes: filters.letterCodes,
     numberRange: filters.numberRange,
@@ -9,9 +12,9 @@ export const submitSearch = (filters: { query: string, letterCodes: Array<string
 };
 
 // eslint-disable-next-line arrow-body-style
-export const changeFilters = (filters: { query: string, letterCodes: Array<string>, numberRange: Array<number> }) => {
+export const submitSearch = (filters: Filters) => {
   return {
-    type: 'CHANGE_FILTERS',
+    type: SUBMIT_SEARCH,
     query: filters.query,
     letterCodes: filters.letterCodes,
     numberRange: filters.numberRange,
